Refactor: sortByAZ :arrow_right: sortBy(sortType)
