+++ tests/client/actions/sortAZ.test.js
@@ -1,8 +1,8 @@
-import { SORT_AZ } from '../../client/actions/sort'
+import { SORT_AZ, sortBy } from '../../../client/actions/sort'
 
 describe('Tests for SortAZ actions', () => {
   it('setSortType returns a sortType', () => {
-    const action = sortByAZ()
+    const action = sortBy(SORT_AZ)
     expect(action.type).toBe(SORT_AZ)
   })
 })
